requirements.txt
cssselect==1.1.0
lxml==4.6.2
python-decouple==3.4
requests==2.25.1
selenium==3.141.0
